⚙️ Move user storage to storage folder
